+++ configuration-parent/configuration-yaml-interpreter/src/main/java/com/github/starnowski/posmulten/configuration/yaml/validation/groups/ValidatorGroupResolver.java
@@ -1,0 +1,12 @@
+package com.github.starnowski.posmulten.configuration.yaml.validation.groups;
+
+import com.github.starnowski.posmulten.configuration.yaml.model.SharedSchemaContextConfiguration;
+
+public interface ValidatorGroupResolver<T> {
+
+    Class<T> resolveForSharedSchemaContextConfiguration(SharedSchemaContextConfiguration sharedSchemaContextConfiguration, ValidatorGroupResolverContext context);
+
+    interface ValidatorGroupResolverContext {
+
+    }
+}